GroupFunctions.jl Documentation

GroupFunctions.jl Documentation

Functions

Functions

GroupFunctions.group_function — Function

group_function(λ::Irrep, tu::YTableau, tv::YTableau)

Return the symbolic group function corresponding to irrep λ and semistandard Young Tableaux (STYT) tu and tv.

Examples:

julia> t = YoungTableau([2,1]); fill!(t, [1,2,3]);
julia> group_function([2,1,0], t, t)

group_function(λ::Irrep, tu::GTPattern, tv::GTPattern)

Return the symbolic group function corresponding to irrep λ and GT patterns tu and tv.

Examples:

julia> t = GTPattern([[2,1,0],[2,1],[2]],[2]);
julia> group_function([2,1,0], t, t)

group_function(λ::Irrep, tu::GTPattern, tv::GTPattern, mat::Array{Complex{Float64},2})

Return the numeric group function, for an SU(n) member mat, corresponding to irrep λ and STYT tu and tv.

Examples:

julia> using RandomMatrices
julia> mat = rand(Haar(2),3)
julia> t = YoungTableau([2,1]); fill!(t, [1,2,3]);
julia> group_function([2,1,0], t, t, mat)

group_function(λ::Irrep, tu::GTPattern, tv::GTPattern, mat::Array{Complex{Float64},2})

Return the numeric group function, for an SU(n) member mat, corresponding to irrep λ and GT patterns tu and tv.

Examples:

julia> using RandomMatrices
julia> mat = rand(Haar(2),3)
julia> t = GTPattern([[2,1,0],[2,1],[2]],[2]);
julia> group_function([2,1,0], t, t, mat)

GroupFunctions.mma_to_julia — Function

mma_to_julia(s::String)

Calcula el polinomio de MMA a uno de Julia usando SymEngine

Examples:

julia > mma_to_julia("x[1, 1])
u_1_1